// frontend/src/generated/email_service.ts
export interface EmailPreviewMessage {
    /**
     * @generated from protobuf field: string id = 1
     */
    id: string;
    /**
     * @generated from protobuf field: string from = 2
     */
    from: string;
    /**
     * @generated from protobuf field: string subject = 3
     */
    subject: string;
    /**
     * @generated from protobuf field: google.protobuf.Timestamp received_at = 4
     */
    receivedAt?: Timestamp;
    /**
     * @generated from protobuf field: bool unread = 5
     */
    unread: boolean;
}
// ...
export interface ListSentResponse {
    /**
     * @generated from protobuf field: repeated com.emailclient.backend.grpc.EmailPreviewMessage emails = 1
     */
    emails: EmailPreviewMessage[];
}
// ...
export interface ListTrashResponse {
    /**
     * @generated from protobuf field: repeated com.emailclient.backend.grpc.EmailPreviewMessage emails = 1
     */
    emails: EmailPreviewMessage[];
}
// ...
class EmailPreviewMessage$Type extends MessageType<EmailPreviewMessage> {
    constructor() {
        super("com.emailclient.backend.grpc.EmailPreviewMessage", [
            { no: 1, name: "id", kind: "scalar", T: 9 /*ScalarType.STRING*/ },
            { no: 2, name: "from", kind: "scalar", T: 9 /*ScalarType.STRING*/ },
            { no: 3, name: "subject", kind: "scalar", T: 9 /*ScalarType.STRING*/ },
            { no: 4, name: "received_at", kind: "message", T: () => Timestamp },
            { no: 5, name: "unread", kind: "scalar", T: 8 /*ScalarType.BOOL*/ }
        ]);
    }
    create(value?: PartialMessage<EmailPreviewMessage>): EmailPreviewMessage {
        const message = globalThis.Object.create((this.messagePrototype!));
        message.id = "";
        message.from = "";
        message.subject = "";
        message.unread = false;
        if (value !== undefined)
            reflectionMergePartial<EmailPreviewMessage>(this, message, value);
        return message;
    }
    internalBinaryRead(reader: IBinaryReader, length: number, options: BinaryReadOptions, target?: EmailPreviewMessage): EmailPreviewMessage {
        let message = target ?? this.create(), end = reader.pos + length;
        while (reader.pos < end) {
            let [fieldNo, wireType] = reader.tag();
            switch (fieldNo) {
                case /* string id */ 1:
                    message.id = reader.string();
                    break;
                case /* string from */ 2:
                    message.from = reader.string();
                    break;
                case /* string subject */ 3:
                    message.subject = reader.string();
                    break;
                case /* google.protobuf.Timestamp received_at */ 4:
                    message.receivedAt = Timestamp.internalBinaryRead(reader, reader.uint32(), options, message.receivedAt);
                    break;
                case /* bool unread */ 5:
                    message.unread = reader.bool();
                    break;
                default:
                    let u = options.readUnknownField;
                    if (u === "throw")
                        throw new globalThis.Error(`Unknown field ${fieldNo} (wire type ${wireType}) for ${this.typeName}`);
                    let d = reader.skip(wireType);
                    if (u !== false)
                        (u === true ? UnknownFieldHandler.onRead : u)(this.typeName, message, fieldNo, wireType, d);
            }
        }
        return message;
    }
    internalBinaryWrite(message: EmailPreviewMessage, writer: IBinaryWriter, options: BinaryWriteOptions): IBinaryWriter {
        /* string id = 1; */
        if (message.id !== "")
            writer.tag(1, WireType.LengthDelimited).string(message.id);
        /* string from = 2; */
        if (message.from !== "")
            writer.tag(2, WireType.LengthDelimited).string(message.from);
        /* string subject = 3; */
        if (message.subject !== "")
            writer.tag(3, WireType.LengthDelimited).string(message.subject);
        /* google.protobuf.Timestamp received_at = 4; */
        if (message.receivedAt)
            Timestamp.internalBinaryWrite(message.receivedAt, writer.tag(4, WireType.LengthDelimited).fork(), options).join();
        /* bool unread = 5; */
        if (message.unread !== false)
            writer.tag(5, WireType.Varint).bool(message.unread);
        let u = options.writeUnknownFields;
        if (u !== false)
            (u == true ? UnknownFieldHandler.onWrite : u)(this.typeName, message, writer);
        return writer;
    }
}
/**
 * @generated MessageType for protobuf message com.emailclient.backend.grpc.EmailPreviewMessage
 */
export const EmailPreviewMessage = new EmailPreviewMessage$Type();
// ...
class ListSentResponse$Type extends MessageType<ListSentResponse> {
    constructor() {
        super("com.emailclient.backend.grpc.ListSentResponse", [
            { no: 1, name: "emails", kind: "message", repeat: 2 /*RepeatType.UNPACKED*/, T: () => EmailPreviewMessage }
        ]);
    }
    create(value?: PartialMessage<ListSentResponse>): ListSentResponse {
        const message = globalThis.Object.create((this.messagePrototype!));
        message.emails = [];
        if (value !== undefined)
            reflectionMergePartial<ListSentResponse>(this, message, value);
        return message;
    }
    internalBinaryRead(reader: IBinaryReader, length: number, options: BinaryReadOptions, target?: ListSentResponse): ListSentResponse {
        let message = target ?? this.create(), end = reader.pos + length;
        while (reader.pos < end) {
            let [fieldNo, wireType] = reader.tag();
            switch (fieldNo) {
                case /* repeated com.emailclient.backend.grpc.EmailPreviewMessage emails */ 1:
                    message.emails.push(EmailPreviewMessage.internalBinaryRead(reader, reader.uint32(), options));
                    break;
                default:
                    let u = options.readUnknownField;
                    if (u === "throw")
                        throw new globalThis.Error(`Unknown field ${fieldNo} (wire type ${wireType}) for ${this.typeName}`);
                    let d = reader.skip(wireType);
                    if (u !== false)
                        (u === true ? UnknownFieldHandler.onRead : u)(this.typeName, message, fieldNo, wireType, d);
            }
        }
        return message;
    }
    internalBinaryWrite(message: ListSentResponse, writer: IBinaryWriter, options: BinaryWriteOptions): IBinaryWriter {
        /* repeated com.emailclient.backend.grpc.EmailPreviewMessage emails = 1; */
        for (let i = 0; i < message.emails.length; i++)
            EmailPreviewMessage.internalBinaryWrite(message.emails[i], writer.tag(1, WireType.LengthDelimited).fork(), options).join();
        let u = options.writeUnknownFields;
        if (u !== false)
            (u == true ? UnknownFieldHandler.onWrite : u)(this.typeName, message, writer);
        return writer;
    }
}
/**
 * @generated MessageType for protobuf message com.emailclient.backend.grpc.ListSentResponse
 */
export const ListSentResponse = new ListSentResponse$Type();
// ...
class ListTrashResponse$Type extends MessageType<ListTrashResponse> {
    constructor() {
        super("com.emailclient.backend.grpc.ListTrashResponse", [
            { no: 1, name: "emails", kind: "message", repeat: 2 /*RepeatType.UNPACKED*/, T: () => EmailPreviewMessage }
        ]);
    }
    create(value?: PartialMessage<ListTrashResponse>): ListTrashResponse {
        const message = globalThis.Object.create((this.messagePrototype!));
        message.emails = [];
        if (value !== undefined)
            reflectionMergePartial<ListTrashResponse>(this, message, value);
        return message;
    }
    internalBinaryRead(reader: IBinaryReader, length: number, options: BinaryReadOptions, target?: ListTrashResponse): ListTrashResponse {
        let message = target ?? this.create(), end = reader.pos + length;
        while (reader.pos < end) {
            let [fieldNo, wireType] = reader.tag();
            switch (fieldNo) {
                case /* repeated com.emailclient.backend.grpc.EmailPreviewMessage emails */ 1:
                    message.emails.push(EmailPreviewMessage.internalBinaryRead(reader, reader.uint32(), options));
                    break;
                default:
                    let u = options.readUnknownField;
                    if (u === "throw")
                        throw new globalThis.Error(`Unknown field ${fieldNo} (wire type ${wireType}) for ${this.typeName}`);
                    let d = reader.skip(wireType);
                    if (u !== false)
                        (u === true ? UnknownFieldHandler.onRead : u)(this.typeName, message, fieldNo, wireType, d);
            }
        }
        return message;
    }
    internalBinaryWrite(message: ListTrashResponse, writer: IBinaryWriter, options: BinaryWriteOptions): IBinaryWriter {
        /* repeated com.emailclient.backend.grpc.EmailPreviewMessage emails = 1; */
        for (let i = 0; i < message.emails.length; i++)
            EmailPreviewMessage.internalBinaryWrite(message.emails[i], writer.tag(1, WireType.LengthDelimited).fork(), options).join();
        let u = options.writeUnknownFields;
        if (u !== false)
            (u == true ? UnknownFieldHandler.onWrite : u)(this.typeName, message, writer);
        return writer;
    }
}
/**
 * @generated MessageType for protobuf message com.emailclient.backend.grpc.ListTrashResponse
 */
export const ListTrashResponse = new ListTrashResponse$Type();